requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
pydantic